requirements.txt
# Core libraries
langchain
llama-index
haystack


# for a-backend server (optional but recommended)
fastapi
uvicorn


# Additional tools
requests
python-dotenv


# Optional AI agent frameworks
# Use Git URLs or manual installs
autogpt
babyagipip 
